# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Tourist",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "localhost",
            "variables": {
                "schema": {
                    "enum": [
                        "https",
                        "http"
                    ],
                    "default": "http"
                }
            }
        }
    ],
    "paths": {
...
        "/sitios/list": {
            "get": {
                "tags": [
                    "Sitios turísticos"
                ],
                "summary": "Listado de sitios turisticos",
                "operationId": "App\\Http\\Controllers\\API\\SitiosController::list_sitios",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/sitios/add": {
            "post": {
                "tags": [
                    "Sitios turísticos"
                ],
                "summary": "Adiciona un sitio turístico",
                "operationId": "App\\Http\\Controllers\\API\\SitiosController::add_sitio",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string"
                                    },
                                    "ubicacion": {
                                        "type": "string"
                                    },
                                    "distancia": {
                                        "type": "float"
                                    },
                                    "tiempo": {
                                        "type": "int"
                                    },
                                    "poblacion": {
                                        "type": "int"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "nombre": "nombre del sitio turistico",
                                    "ubicacion": "localización",
                                    "distancia": 0,
                                    "tiempo": 0,
                                    "poblacion": 1000
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/sitios/uploadphoto/{idsitio}": {
            "post": {
                "tags": [
                    "Sitios turísticos"
                ],
                "summary": "Sube al servidor la foto del sitio",
                "operationId": "App\\Http\\Controllers\\API\\SitiosController::subir_imagen",
                "parameters": [
                    {
                        "name": "idsitio",
                        "in": "path",
                        "description": "Identificador del sitio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image_sitio": {
                                        "type": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/sitios/calificar/{idsitio}/{idimagen}/{calificacion}": {
            "post": {
                "tags": [
                    "Sitios turísticos"
                ],
                "summary": "Califica una imagen",
                "operationId": "App\\Http\\Controllers\\API\\SitiosController::calificar_imagen",
                "parameters": [
                    {
                        "name": "idsitio",
                        "in": "path",
                        "description": "Identificador del sitio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "idimagen",
                        "in": "path",
                        "description": "Identificador de la imagen",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "calificacion",
                        "in": "path",
                        "description": "Calificacion",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/sitios/imagen_position/{idimagen}": {
            "post": {
                "tags": [
                    "Sitios turísticos"
                ],
                "summary": "Salvar posicion imagen",
                "operationId": "App\\Http\\Controllers\\API\\SitiosController::save_position",
                "parameters": [
                    {
                        "name": "idimagen",
                        "in": "path",
                        "description": "Identificador de la imagen",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lat": {
                                        "type": "string"
                                    },
                                    "lon": {
                                        "type": "string"
                                    },
                                    "zoom": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "lat": "lat",
                                    "lon": "lon",
                                    "zoom": "zoom"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/sitios/delete/{idsitio}": {
            "delete": {
                "tags": [
                    "Sitios turísticos"
                ],
                "summary": "Elimina un sitio turístico",
                "operationId": "App\\Http\\Controllers\\API\\SitiosController::delele_sitio",
                "parameters": [
                    {
                        "name": "idsitio",
                        "in": "path",
                        "description": "Identificador del sitio",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Failed"
                    }
                },
                "deprecated": false,
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "login",
                "name": "token",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}